src/requirements.txt
python-dotenv==1.0.1
tzlocal==5.2
scrapy==2.11.2
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
